# task-api-openapi.yaml
openapi: 3.0.3
info:
  title: Task API
  version: 1.0.0
  description: Spring Boot in-memory Task Handler API
servers:
  - url: http://localhost:8080

paths:
  /tasks:
    get:
      summary: Get all tasks
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"

    post:
      summary: Create a task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskCreateUpdateRequest"
      responses:
        "201":
          description: Task created
          headers:
            Location:
              description: URL of created task
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /tasks/{id}:
    get:
      summary: Get task by ID
      parameters:
        - $ref: "#/components/parameters/TaskId"
      responses:
        "200":
          description: Task found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          description: Task not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    put:
      summary: Update task by ID
      parameters:
        - $ref: "#/components/parameters/TaskId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskCreateUpdateRequest"
      responses:
        "200":
          description: Task updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Task not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      summary: Delete task by ID
      parameters:
        - $ref: "#/components/parameters/TaskId"
      responses:
        "204":
          description: Task deleted (no content)
        "404":
          description: Task not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /tasks/info:
    get:
      summary: Get app info
      responses:
        "200":
          description: App metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppInfo"

components:
  parameters:
    TaskId:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64

  schemas:
    Task:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        status:
          type: string
          enum: [OPEN, IN_PROGRESS, DONE]
      required: [id, title, status]

    TaskCreateUpdateRequest:
      type: object
      properties:
        title:
          type: string
        status:
          type: string
          enum: [OPEN, IN_PROGRESS, DONE]
      required: [title, status]

    AppInfo:
      type: object
      properties:
        appVersion:
          type: string
        appMode:
          type: string
        port:
          type: string
      required: [appVersion, appMode, port]

    ErrorResponse:
      type: object
      properties:
        status:
          type: integer
        error:
          type: string
        message:
          type: string
      required: [status, error, message]